Fix for #1722
Reported was that padplus/padminus doesn't do a zoom in the NLA or Action
windows. Was missing since very beginning! :)
Found conflict with the new NLA strip move up/down, but using the pad
keys for that is a bit too inconsistant. I've now made it using the
PageUp/PageDown and normal plus/minus keys... Hos could check on this
though!