Bugfix #6624:
When the databrowser was used for displaying available IPO-blocks, when called from the IPO-editor header, only Object IPO blocks were shown. This bug has apparently been around for quite a few releases, probably due to few people using
this feature (it only occurs on files with > 60 IPO-blocks of one type, and only
when you try to reassign ipo-blocks too).