-> Face split tools
Two simple face split tools, 'connect verts' and 'connect edges'.
In the vertex connect tool, for any face in the mesh, if it has 
two selected vertices, it will connect them by splitting the face 
in two and adding an edge. Similarly in the edge connect tool, 
for any face in the mesh with two selected edges it will split them
in half and then split the face by connecting the two new vertices.


Note that these are mostly simple examples and to give people something
to play with. In reality there should not be the restriction of just
connecting two verts/edges in a face at a time.