== Sequencer ==
Fixed a bug with ibuf caching on startstill / endstill.

The new blend modes happened to force start and endstill to be rendered
over and over again. This could get very annoying especially on scene
strips.

We therefore now cache the original start or endstill ibuf seperately
and copy on demand.