Bugfix #8948: Transformation constraint - destination rotation is in radians
Unfortunately, my previous fixes here to get gears working resulted in this problem. The flip side of that was that gears worked for quite a few cases. 

Alas, it is not to be... The bug is now fixed, but gears no longer work any more. Sadly, there is not much I can do about this.