AnimSys2: Per-Segment Interpolation of IPO-Curves
It is now possible to define the interpolation that will be used to get from a keyframe until the next one is encountered.
With an Ipo-Curve in EditMode, only the selected keyframes will get the specified interpolation mode set. 
Out of EditMode, the interpolation mode will be applied to all the keyframes of an IPO-curve that is selected.

- When inserting keyframes for curves that have multiple types of interpolation in use, the interpolation mode of the keyframe immediately before the added one is used.
- Old files get converted automatically
- Subversion has been bumped up to 2, so that old files get converted properly.