AnimSys2: IPO Editor Handle Drawing
As an experiment, handle vertices in IPO Editor are now drawn as unfilled circles. This is more in line with what other packages are doing, and is aimed at making the handles more visibly different from the actual keyframes. 

Comments are welcome on whether these contribute more visual clutter than worth, any performance hits on more intensive shots, or OpenGL incompatabilities (there shouldn't be issues, as other parts of Blender does this too).