The conversions functions are now operator-afied.
Also tried to make the flags play nice with each other
again, since that was messing things up.  I'm a
little confused on the design; I thought that flags
from the Edit****->f variables would go in
header->flag, along with the flags that live in
other variables.

I made the two temp flags (flag layer flag, not
header->flag) be at the 15th and 16th bit, since
otherwise they could probably interfere with
operator flags.  I also reverted the change of
header->flag to a short, and the changes of the
bmesh-specific flags.