Patch [#19293] NearSensor: Changing distance and resetDistance from Python did not take effect immediately from Frank Papenmeier (hodka)  --- from the submission Changing the distance and resetDistance Properties of a NearSensor from Python did only take effect, once there was a collision within the old radius.
Reason:
The radius of the PHY_IPhysicsController was not set to the new values.

Solution:
It is set at the end of the CheckResetDistance function. I moved the relevant code snippet from the Evaluate function
into a new function called SetPhysCtrlRadius in order to avoid duplicating code doing the same thing.