Copy Constraints Operator the C code to copy the constraint was already there (same as 2.49) so this operator simply exposes it. I'm using the name Copy Constraints to Others based on what we have in material operators (e.g. Copy Material to Others).
If we bring back the whole Copy Attributes menu (old Ctrl+C) having the individual operators is still important for scripts. Nevertheless I'm not comfortable with the name (to Others sounds unnecessary verbose) but so far having the functionality back is important (request from Caju Studio). I'll be glad to make any change later if needed.
(no shortcut for that, to use it go to userpref->keymap and assign a shortcut for this operator or add it to the toolshelf)