Fix [#22097] missing panels in texture tab
Made texture/texture slot context a bit less flaky when dealing with active material and 
texture nodes inside a node material in the node editor. Now if the active material has 
nodes enabled, and there are no active material/texture nodes inside it, nothing will be 
shown in the texture properties (similar to 2.49).