=Hair Dampening and Collision=
Some durian features; added a new dampening mode to hair (and cloth, since
they shared the same dynamics code), that blends velocities with what the velocity
would be if the hair or cloth was a stiff rigid body.  It's called "flexibility dampening"
for now.  Note that the code doesn't go through the solver directly, but this *is* an implicit
solver, and integrating this sort of outside stuff is the whole point of implicit systems.

Plus it works better for solving durian issues as quickly as possible (jahka: pin stiffness for hair is totally unusable btw).

Also added collision for hair, extending the existing collision code.  I have more reliable
code then this, but it'd b a bit of work to have it co-exist with the existing collision 
code (the other code replaces all of the collision code, so I'd have to rework it to coexist instead).