Partial fix for [#22574] Logic Panel missing buttons (when selecting multiple objects) This commit allows you to see the Logic Bricks for multiple objects at once. It still will only add s/c/a for the active object.
@Matt,
currently "LOGIC_OT_controller_add" uses the active object. That's good for the operator to work in scripts, however for the UI we need something different.
Ideally I would like to pass the object as an (optional) parameter to the operator. Not sure if it's possible.

The solution in 2.49 looks too "2.50 incompatible". In there ob->scaflag is set to be retrieve later by "do_logic_buts". Smart but too hacky imho.