== Wax & Clay Brushes == The old clay brush has been renamed 'Wax', this brush fills in crevices below the offset plane and does not effect brushes above it.
A new Clay brush has been created that is like 'Flatten' with a high plane offset.  It brings things up to the offset plane while lowering things above it.

On top of that, the default offset for clay and wax has been lowered to half the brush radius and the plane offset slider now only goes from -0.5 to 0.5.  This has the effect of flattening clay and wax and makes them less like draw

(so please, no more "clay is just like draw" comments :P)