== Symmetry Feather ==
Symmetry feather reduces the strength of the brush if it overlaps with other symmetrical instances of the brush.  Works in mirror and radial modes.

Rearranged UI because symmetry seems to have enough options to deserve its own panel now.

Made 'lock axis' into a set of toggle buttons so if takes up less space.