* Fix for relative paths to style module files.
When a .blend file is saved, the "Remap Relative" option allows users
to keep relative paths valid even if the .blend file is saved into a
different directory.  Now the Remap Relative option takes care of relative
paths to style module files.  In addition, the following operations work
as expected with respect to style modules:
- File >> External Data >> Make All Path Relative
- File >> External Data >> Make All Path Absolute
- File >> External Data >> Report Missing Files

* Indentation fix in the UI code (no functionality changes).