New option for taking account of face smoothness in view map calculation.
With this option disabled, the smoothness of faces is ignored and
jagged feature edges (when seen from a three-dimensional viewpoint)
are generated.  If the new option is enabled, face smoothness is taken
into account during the view map calculation, resulting in smooth
feature edges in 3D.  This generally leads to more stable strokes, and
also permits more stylization possibilities relying on the feature
edge smoothness in 3D.  The downside of smooth feature edges is that
the visibility is mathematically not well-defined and may cause some
artefact in rendering results.  The new option is thus intended to
allow users to try both jagged and smooth feature edges and see which
ones would fit individual application needs.  This option is disabled
by default.