Fix [#26474] double_sided wrong in COLLADA Export reported by Juan Linietsky
The export was not really wrong, but Maya expects their profile to be present inside <geometry>. Added this for mesh with ME_TWOSIDED flag set.

Ideally this will change in the future to be controllable through options to the exporter (like current Google Earth and 3DS Max extensions for <double_sided>).