New composite engine.
committed in order to start the migration of the nodes.
have some issues with some changes in trunk based on the UI event handlers.
sometimes an exception happens in the malloc of a UI-control.
need to discuss what the issue is.

OpenCL part is disabled right now. will be enabled in a folowing commit.

Jeroen & Monique.
 - At Mind -