Distance Matte node - Tile this matte operation is outputting RGBA (in oppose to only the A value as some other mattes). The reason for this is to avoid and extra operation to get the main output ready (which is always needed for the preview so we may as well get it done).
Instead of using the SetAlphaOperation I'm using SeparateChannel to extract the alpha from the output
of the Distance Matte operation.

Jeroen, I think it's simpler if I go ahead and commit the nodes if I think they are good.
If there are problems it'll be simpler to fix in svn since they should be small.