Normalize Node - Tile The implementation is simple. Fancy alternatives to gather the minimum and maximum through recursive loop in the tiles proved inefficient (Jeroen Bakker's  tests)
I'm also using the generic NodeTwoFloats to store the temporary values calculated in the loop.
And different from Blender trunk I'm not checking to see if the value is between the accepted Z range
(e.g. (*val <= BLENDER_ZMAX)). I can't see why that would be necessary. To avoid overflow? If so this be handled (aka clamped) outside the nodes in a generic way.

Credits to Jeroen Bakker for the solution in using mutex (copy pasted from his ToneMap code) and review