UV transform correction tool ============================
*Add struct UVTransCorrection for the tool and cleanup code for the
struct.
*After some thought it looks like I cannot use unwrap for this tool
since it will likely invalidate any custom uv layout made by the artist
and replace it with the position from the unwrapper. So I will do it
with the second method, by modifying the uvs based on the ratios of edge
lengths. This will probably require some mathematical correction that I
will have to work on.