Tomato Cycles: resolve textures limit when using CPU render
Remove limit for number of textures when using CPU rendering,
so huge scenes from Mango could be rendered properly.

Re-arranged float and bytes textures for this, so now float
textures goes before byte. This is needed to make alignment
stuff easier for variable number of textures.

Also replace code-generated CPU kernel structure's members
with an array, which makes it easier to access textures
without adding bunch of cases in switch.

Should be no functional changes when rendering with GPU
or OpenCL devices.