Isomap Unwrapper ================== * First working version, however:
1) It's extremely slow (finding geodesic distances between all vertices
on mesh is O^4 expensive!) It can be optimized away, however there is
problem number...
2) The result is not that good when only using mesh connectivity from
edges. I may have to look at an alternative to calculate distances.

However, reading some more papers it looks like this is quite an
expensive algorithm. Coupled with the fact that it can't support pinned
UVs this will probably as far as this will go since there is no time to
investigate other methods. I will try to see if result can be improved a
bit though.