Fix #33345: Crash when using bpy.ops.sculpt.brush_stroke
It was kind of a regression in behavior in svn rev46862 which made it so
blender crashes if stroke is done from the script.

It should bring back the behavior back and made it so blender doesn't crash,
however it's probably not full fix and some further work is needed to make
call of stroke operator usable from the addon.