Motion tracking: automatic keyframe selection
This commit contains implementation of automatic keyframe selection algorithm
based on Pollefeys's criteria (F-GRIC is smaller than H-GRIC and correspondence
ratio is more then 90%). It is implemented as a part of simple pipeline and
returns vector of keyframe images for a given Tracks structure.

For simple pipeline reconstruction two best keyframes are selecting from list
of all possible keyframes. Criteria for this selection is reprojection error
of solution from two candidate keyfames.

In Blender side added an option in Solve panel to enable Keyframe Selection.
If this option enabled, libmv will detect two best keyframes and use them for
solution. This keyframes will be set back to the interface when solution is
done.