Bug fix, irc submitted:
On saving a first .blend after startup, the file would load back as if it was not
saved (showing no name in header). 

The whole FILEFLAGS and G.relabase_valid and G.file_saved etc is messy.
This commit fixes issues, but only adds more mess :)

Will discuss a nicer implementation of all of this.