Add first version of pressure influence for mask textures based on feedback from David Revoy:
http://www.davidrevoy.com/article107/textured-brush-in-floss-digital-

There are two methods, cutoff and ramp.

Cutoff simply selects between 0 and 1 based on stylus pressure while
ramp distributes the mask effect above the pressure value.

Too many words, here is a screenshot with the effect:

http://www.pasteall.org/pic/show.php?id=53865

From top to bottom:
* Strength pressure influence ON:
     - No mask influence
     - Cutoff mask influence
     - Ramp mask influence
* Strength pressure influence OFF:
     - No mask influence
     - Cutoff mask influence
     - Ramp mask influence
     
The brush masking is still in need of attention because it behaves
erratically, still it's nice for an zero.alpha commit ;)