Change masking of 2d painting to be similar to 3d masking.
This fixes inconsistent behaviour where opacity masking is done using
only the curve and alpha offset when view mask mapping is used in 2d
painting. The key notion here is that masking converges to the maximum
value, which should be texmask*brush_alpha here. This happens already
for projective texturing.

Unfortunately this requires an extra allocation of a maximum mask value
imbuf during a stroke which introduces some performance regression when
using mask inages. This first version of the fix does not yet introduce
partial update of this maximum mask buffer like we do for colour image
buffers, and even though this bug exists in trunk as well, I decided to
only commit in this branch because it is not too obvious and because
there is some performance regression when using mask brushes.

Also, this commit may break pressure masking because I don't have a
tablet right now to test. Don't panic if this breaks it, I will fix it
when I get my hands on it again. However if this is not the case, this
commit should also fix the 2d behaviour of ramp style pressure masking.