Drag dot strokes for image painting.
This was mostly working with undo code.

There is a tradeoff between doing undo traversal for images vs undo
initialiatio time:

The more area the user paints on, the more undo tiles are added to the
system. That makes restoring of the image slower after each step.
A way to optimize this is to completely delete the undo tiles after each
step. This also will make the undo steps use less memory.

However given that for a realistic, non-user-testing use case, the dabs
will be localized, better optimize the initialization case. In
projective texturing, which is the most performance hungry, indeed this
helps with performance, provided the use case is sane. I left the
clearing code ifdefed out just in case.

Also exposed anchored stroke options. They work with texture coordinates
well but not with paint brush size yet.