Line style strokes.
Access from stroke panel or by pressing A.
To use press on first spot and drag to end spot. A line is displayed to
show the path of drawing. Pressure is disabled in this mode. Rationale
is that you can't get meaningful pressure values at the endpoints of the
line only.

It would be nice if we had some icons for the stroke modes.

Also, airbrushes now only trigger on timer events and cursor is always
refreshed when a stroke is active. Based on a patch by Jason Wilkins,
thanks!