Enabled GLEW_MX and got it working with GHOST_ContextWGL
GLEW_MX puts each set of GL entry points into its own independent context.  This should reduce the possibility of WGL and GL getting mismatched.

It should also allow desktop and embedded GL to live in the same program at the same time, which isn't really the end goal, but if multiple versions of OpenGL could run in the same instance of Blender it would mean that GL contexts are well managed and encapsulated.

The biggest chore is probably sharing textures between alien contexts, however, I believe that problem can be solved by the same mechanism that will be needed to allow Blender to recover from power management events on mobile devices (and probably also the desktop when using ES).

Not going to go too far down that road now, but this is the foundation.