Initial code layout for plane track deform node
Idea of this is:
- User selects which plane track to use (for this he
  need to select movie clip datablock, object and track
  names).
- Node gets an image and mask inputs (both are optional).
- Node outputs:
  * Input image warped into the plane.
  * Input mask warped by the plane.
  * Plane, rasterized to a mask.

Warping image is done by computing reverse bilinear
coordinates, and getting pixel from corresponding
position.

This requires some tricks with downsampling to make warped
image looks smooth.

Currently compositor doesn't support downsampling, so we
needed to implement our own operation for this.

Currently idea is dead simple: value of output pixel equals
to an average of neighborhood of corresponding pixel in input
image. Size of neighborhood is defined by ratio between input
and output resolutions.

This operation doesn't give perfect results and works only
for downsampling. But it's totally internal operation not
exposed to the interface, so it's easy to replace it with
smarter bi-directional sampler with nicer filtering.

Limitations:
- Node currently only warps image and outputs mask created
  out of plane, warping input mask is not implemented yet.
- Image warping doesn't report proper depending area of
  interest yet, meaning interactivity might be not so much
  great,
- There's no anti-aliasing applied on the edges of warped
  image and plane mask, so they look really sharp at this
  moment.