Replace crappy custom sampler with EWA one
Our sampler could be useful for more general usage,
like simpler downsampling in scale node, but it needs
to be improved before it's really useful.

It was giving lots of jittering artifacts which isn't
good for motion tracking. So now image wrap operation
used EWA filtering.

For now, it uses copy-pasted dx/dy calculation which
came from MapUV operation. Ideally we need to get
rid of duplicated code, but it's not so trivial for
now because of UV coordinates are calculating in
different ways. Not a big deal to have duplicated
code for a while.