Added check for last datamask used for derivedRender calculation
That's perhaps not too much difference for now, because it seems
all places are using the same data mask for derived render, but
better be safe and robust here than sorry.

Alternative would be to remove dataMask from API here and always
use hard-coded mask in DerivedMesh.c for derivedRender.