rigidbody: Add effector weights for individual rigid bodies
There's a need for more fine grained control over how individual rigid
bodies are affected by force fields.
Having effector weights for both the world and bodies might be too much.
If so, we could drop it from the world and copy the world settings to
the body setting for old files.