Add first working code for sharpen tool in projective texturing.
To use, select the Soften tool and press Ctrl while painting.

Notes:

* Disabled masking for the soften tool, it did some weird effects, such
as brush curve having weird effects (full curve would not work, for
instance).

* Unsharp masking requires a better filter to address changes in color
in the diagonal direction. Currently it shares the filter used by the
soften brush, which is similar to a simple cross filter.

* Threshhold will be added as well, will only add shariness if edge
detection value exceeds it. Will keep other parts of the image nice and
clean.